pyproject.toml
[project]
name = "nfe-parser-gemma3n"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-ollama>=0.3.6",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.116.1",
]
